+++ CHANGELOG.md
@@ -1,12 +1,20 @@
-## 0.2.1
+## 0.3.0
+
+- **BREAKING CHANGE**: The `arguments` argument to `ProcessManager.spawn` is
+  now positional (not named) and required. This makes it more similar to the
+  built-in `Process.start`, and easier to use as a drop in replacement:
+
+```dart
+processManager.spawn('dart', ['--version']);
+```
 
 - Fixed a bug where processes created from `ProcessManager.spawn` could not
   have their `stdout`/`stderr` read through their respective getters (a runtime
   error was always thrown).
 
-- Added `ProcessMangaer.background`, which does not forward `stdin`.
+- Added `ProcessMangaer#spawnBackground`, which does not forward `stdin`.
 
-- Added `ProcessManager.headless`, which does not forward any standard I/O.
+- Added `ProcessManager#spawnDetached`, which does not forward any I/O.
 
 ## 0.2.0